requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
redis==5.0.8
rq==1.16.2
pypdfium2==4.30.0
Pillow==10.4.0
zxing-cpp==2.2.0
python-multipart==0.0.9
